Update cod-hgy7ss_cy8th8-d1e58.xml
I added a note: <!--It is really a spelling mistake, but also a grammatical mistake made by the scribe. It is possible that the scribe had in mind that the word "reperio" belongs to the fourth conjugation, just like "audio", whose past form is "audivi". However, "reperio" actually belongs to the mixed conjugation and is conjugated in the past like third-conjugation verbs, just like "lego", whose past form is "legi".→